requirements.txt
faiss-cpu==1.9.0
sentence-transformers==3.3.1
rank-bm25==0.2.2
tiktoken==0.8.0
anthropic==0.40.0
openai==1.56.0
pypdf==5.1.0
python-docx==1.1.2
pyyaml==6.0.2
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.3.4
pytest-asyncio==0.24.0
streamlit>=1.40.0
cryptography==41.0.7
